Split app module to lazy load calculators
