+++ src/app/app-routing.module.ts
@@ -1,14 +1,20 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { UsableRoutes } from './usable-routes';
-import { PageIntroComponent } from './pages/page-intro/page-intro.component';
-import { PageCalculatorsComponent } from './pages/page-calculators/page-calculators.component';
 
 const routes: Routes = [
-  {path: UsableRoutes.Intro, component: PageIntroComponent},
+  {
+    path: UsableRoutes.Intro,
+    loadChildren: () => import('./article/article.module').then(m => m.ArticleModule)
+  },
 
-  {path: UsableRoutes.DvPlanner, component: PageCalculatorsComponent},
-  {path: UsableRoutes.SignalCheck, component: PageCalculatorsComponent},
+  {
+    path: 'calculators',
+    loadChildren: () => import('./calculators/calculators.module').then(m => m.CalculatorsModule)
+  },
+
+  {path: UsableRoutes.DvPlanner, redirectTo: `calculators/${UsableRoutes.DvPlanner}`},
+  {path: UsableRoutes.SignalCheck, redirectTo: `calculators/${UsableRoutes.SignalCheck}`},
 
   {path: '**', redirectTo: UsableRoutes.Intro},
 ];
