+++ src/app/article/page-intro/page-intro.component.scss
