+++ src/app/calculators/components/page-signal-check/page-signal-check.component.scss
