+++ src/app/calculators/components/page-signal-check/page-signal-check.component.spec.ts
@@ -1,16 +1,16 @@
 import { PageSignalCheckComponent } from './page-signal-check.component';
 import { MockBuilder, MockInstance, MockRender } from 'ng-mocks';
-import { AppModule } from '../../app.module';
-import { StateService } from '../../services/state.service';
+import { AppModule } from '../../../app.module';
+import { StateService } from '../../../services/state.service';
 import { BehaviorSubject, EMPTY, of } from 'rxjs';
-import { HudService } from '../../services/hud.service';
+import { HudService } from '../../../services/hud.service';
 import { ineeda } from 'ineeda';
-import { SpaceObject } from '../../common/domain/space-objects/space-object';
-import { SpaceObjectService } from '../../services/space-object.service';
+import { SpaceObject } from '../../../common/domain/space-objects/space-object';
+import { SpaceObjectService } from '../../../services/space-object.service';
 import { MatDialog } from '@angular/material/dialog';
-import { Craft } from '../../common/domain/space-objects/craft';
-import { CraftDetailsDialogComponent } from '../../overlays/craft-details-dialog/craft-details-dialog.component';
-import { UniverseMapComponent } from '../../components/universe-map/universe-map.component';
+import { Craft } from '../../../common/domain/space-objects/craft';
+import { CraftDetailsDialogComponent } from '../../../overlays/craft-details-dialog/craft-details-dialog.component';
+import { UniverseMapComponent } from '../../../components/universe-map/universe-map.component';
 import createSpy = jasmine.createSpy;
 import objectContaining = jasmine.objectContaining;
 import arrayContaining = jasmine.arrayContaining;
