+++ src/app/overlays/account-dialog/account-dialog.component.html
@@ -1,2 +1,2 @@
-<!--Isolates account details from dialog, preventing circular dependency to state -> auth -> account-->
+<!--  Isolates account details from dialog, preventing circular dependency to state -> auth -> account  -->
 <cp-account-details></cp-account-details>
