+++ src/index.html
@@ -3,6 +3,7 @@
 <head>
   <meta charset="utf-8">
   <title>KSP Visual Calculator</title>
+
   <meta name="description"
         content="Online delta-v calculator for Kerbal Space Program that visualizes and solves problems like the dv costs for a mission, or design an effective communications network." />
   <meta name="keywords"
@@ -23,12 +24,6 @@
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
 </head>
 <body class="mat-typography" style="background-color: #222">
-
-<div style="display: none">
-  <!--  This script creates an iFrame, which in turns gets moved into the base-panel-component.-->
-  <script type="application/javascript" data-idzone="4509002" src="https://a.exdynsrv.com/nativeads-v2.js"></script>
-</div>
-
 <cp-root>
   <style>
     .aspect-ratio-layout {
